user login and some page resistrictions added

+++ base/views.py
@@ -1,6 +1,13 @@
 from django.shortcuts import render, redirect
+from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib import messages
+
 
 # rooms = [
 #     {
@@ -27,11 +34,19 @@ def home(request):
     if query == '':
         rooms = Room.objects.all()
     else:
-        rooms = Room.objects.filter(topic__name__contains = query)
+        # search by topic name and any value from that key word
+        # rooms = Room.objects.filter(topic__name__contains = query)
+        rooms = Room.objects.filter(
+            Q(topic__name__contains = query) | Q(name__icontains=query) | Q(description__icontains=query)
+            )
 
     topic  = Topic.objects.all()
 
-    context = {'rooms': rooms, 'topic': topic}
+
+
+    room_count = rooms.count()
+
+    context = {'rooms': rooms, 'topic': topic, 'room_count': room_count}
     return render(request, 'base/home.html', context)
 
 
@@ -45,7 +60,7 @@ def room(request, pagekey):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
 
@@ -58,9 +73,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request, pagekey):
     room = Room.objects.get(id=pagekey)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse(f"This is {room.host}'s room you can't edit this")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room )
@@ -72,10 +91,42 @@ def updateRoom(request, pagekey):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pagekey):
     room = Room.objects.get(id=pagekey)
+
+    if request.user != room.host:
+        return HttpResponse(f"This is {room.host}'s room you can't delete this")
+
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     context = {'obj': room}
     return render(request, 'base/delete.html', context)
+
+
+
+def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        user = authenticate(request, username=username, password=password) 
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.warning(request, 'Invalid username or password')
+            
+    return render(request, 'base/login.html')
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
